requirements.txt
streamlit
PyMuPDF
langdetect
Pillow
